re-use existing co-proc for child link table
